massive refactor, fixing Lua API and node IO
Make Lua API actually reflect the Rust API

We don't actually need to make this super simple.  We can make simpler
abstractions on top.

Graph now has fully-featured IO, with as many streams as it wants.

This finally allows a graph to be a subgraph.  It is completely
reasonable to have graphs as sub-effects inside larger graphs now.