rewrite Node trait around inner mutability
I don't like this, but it's necessary for dynamic trait upcasting
to work. There is no way to cast `Rc<RefCell<dyn FrequencyNode>>` to
`Rc<RefCell<dyn Node>>` until that is implemented, but we need this
upcasting for instruments to work, so we can have something like
`FrequencyNode` that we can set as nodes in the internal graph.

Once trait object upcasting is stable, then we can go back to the more
reasonable and sane way of doing things.