remove FrequencyNode, bring back Arc<Mutex<Node>>
FrequencyNode has been factored out and is no longer used.  Instead,
nodes are expected to take the frequency in as a constant value input.

With this change, nodes no longer have to be a plain Arc<Node> for
upcasting purposes (and we may have been able to avoid that anyway with
some creative trait casting, now that I think of it).  We've brought
back the Mutex use of these Nodes.